svn path=/main/branches/prefix/; revision=4684

+++ pym/cache/flat_list.py
@@ -1,5 +1,5 @@
-import fs_template
-import cache_errors
+from cache import fs_template
+from cache import cache_errors
 import errno, os, stat
 
 # store the current key order *here*.
@@ -35,7 +35,8 @@ class database(fs_template.FsBased):
 				raise KeyError(cpv)
 			raise cache_errors.CacheCorruption(cpv, e)
 
-		try:	d["_mtime_"] = os.fstat(myf.fileno()).st_mtime
+		try:
+			d["_mtime_"] = long(os.fstat(myf.fileno()).st_mtime)
 		except OSError, e:	
 			myf.close()
 			raise cache_errors.CacheCorruption(cpv, e)
